+++ src/main/java/com/putoet/day9/Day9.java
@@ -7,12 +7,12 @@ import java.util.List;
 
 public class Day9 {
     public static void main(String[] args) {
-        final List<Integer> intCode = CSV.list("/day9.txt", Integer::parseInt).get(0);
+        final List<Long> intCode = CSV.list("/day9.txt", Long::parseLong).get(0);
         part(intCode, "part 1", 1L);
         part(intCode, "part 1", 2L);
     }
 
-    private static void part(List<Integer> intCode, String part, long inputValue) {
+    private static void part(List<Long> intCode, String part, long inputValue) {
         final Memory memory = new ExpandableMemory(intCode);
         final IntCodeInputOutputDevice input = new IntCodeInputOutputDevice();
         final IntCodeInputOutputDevice output = new IntCodeInputOutputDevice();
